This is a work in progress.

```
bash scripts/build-stdlib.sh
lit --succinct --show-all --workers 1 ./benchmarks/collections/bench_bitset.mojo
```

Current results off of 1fa1283383867c33fe301f1c169195247d4bd2c6
```
------------------------------------------------------------------
| name                             | met (ms)            | iters |
|-----------------------|-----------------|------|----------------|
| bench_empty_bitset_init[10]      | 0.3123077928870293  | 3824  |
| bench_empty_bitset_init[30]      | 0.319340625         | 3840  |
| bench_empty_bitset_init[50]      | 0.31526022401667103 | 3839  |
| bench_empty_bitset_init[100]     | 0.31201298026998964 | 3852  |
| bench_empty_bitset_init[1000]    | 0.31210426417056686 | 3846  |
| bench_empty_bitset_init[10000]   | 0.31187525933609955 | 3856  |
| bench_empty_bitset_init[100000]  | 0.3125909801876955  | 3836  |
| bench_empty_bitset_init[1000000] | 0.3131273719781648  | 3847  |
| bench_bitset_init_from[1]        | 0.0906035           | 10000 |
| bench_bitset_init_from[2]        | 0.19275289948453606 | 6208  |
| bench_bitset_init_from[4]        | 0.3852121990369181  | 3115  |
| bench_bitset_init_from[8]        | 0.7663017296604739  | 1561  |
| bench_bitset_init_from[16]       | 1.5751675531914895  | 752   |
| bench_bitset_init_from[32]       | 3.159303664921466   | 382   |
| bench_bitset_init_from[64]       | 6.29592670157068    | 191   |
| bench_bitset_init_from[128]      | 13.003436170212767  | 94    |
| bench_bitset_init_from[256]      | 25.557893617021275  | 47    |
| bench_bitset_init_from[512]      | 51.070913043478264  | 23    |
| bench_bitset_init_from[1024]     | 103.353             | 10    |
| bench_bitset_init_from[2048]     | 214.7138            | 5     |
------------------------------------------------------------------
```

Related to:
- https://github.com/modular/modular/pull/4522
- https://github.com/modular/modular/pull/4511