@gryznar 2025-05-13
I don't know if this is a convention, but most prs starts with uppercase